split up proofread saga into multiple modules
