+++ frontend/javascripts/viewer/model/sagas/volume/proofreading/proofread_saga.ts
@@ -1,144 +1,52 @@
-import {
-  getAgglomeratesForSegmentsFromTracingstore,
-  getEdgesForAgglomerateMinCut,
-  getNeighborsForAgglomerateNode,
-  getPositionForSegmentInAgglomerate,
-  type MinCutTargetEdge,
-  type NeighborInfo,
-} from "admin/rest_api";
-import processTaskWithPool from "libs/async/task_pool";
-import { V3 } from "libs/mjs";
-import { NumberLikeMapWrapper } from "libs/number_like_map_wrapper";
 import Toast from "libs/toast";
-import { getAdaptToTypeFunction, isEditableEventTarget, isNumberMap, SoftError } from "libs/utils";
-import window from "libs/window";
-import { uniq } from "lodash-es";
-import isEqual from "lodash-es/isEqual";
-import uniqBy from "lodash-es/uniqBy";
+import { SoftError } from "libs/utils";
 import messages from "messages";
-import { all, call, put, takeEvery } from "typed-redux-saga";
-import type { AdditionalCoordinate, ServerEditableMapping } from "types/api_types";
-import Constants, {
-  MappingStatusEnum,
-  OrthoViews,
-  TreeTypeEnum,
-  type Vector3,
-} from "viewer/constants";
+import { call, put, takeEvery } from "typed-redux-saga";
+import { OrthoViews, TreeTypeEnum } from "viewer/constants";
 import { getSegmentIdForPositionAsync } from "viewer/controller/combinations/volume_handlers";
 import getSceneController from "viewer/controller/scene_controller_provider";
+import { getVisibleSegmentationLayer } from "viewer/model/accessors/dataset_accessor";
 import {
-  isAgglomerateTree,
-  isAnnotationEditableByNonOwners,
-  isConcurrentCollaborationMode,
-  mayEditAnnotation,
-} from "viewer/model/accessors/annotation_accessor";
-import {
-  getLayerByName,
-  getMagInfo,
-  getMappingInfo,
-  getVisibleSegmentationLayer,
-} from "viewer/model/accessors/dataset_accessor";
-import {
-  areGeometriesTransformed,
   enforceSkeletonTracing,
-  findTreeByNodeId,
   getTreeAndNode,
-  getTreeNameForAgglomerateTree,
 } from "viewer/model/accessors/skeletontracing_accessor";
 import { AnnotationTool } from "viewer/model/accessors/tool_accessor";
 import {
   getActiveSegmentationTracing,
   getActiveSegmentationTracingLayer,
   getActiveUnmappedSegmentId,
-  getEditableMappingForVolumeTracingId,
-  getMeshInfoForSegment,
-  getSegmentsForLayer,
-  getVolumeTracingById,
-  isMeshLoaded,
 } from "viewer/model/accessors/volumetracing_accessor";
+import { removeMeshAction } from "viewer/model/actions/annotation_actions";
 import {
-  dispatchMaybeFetchMeshFilesAsync,
-  removeMeshAction,
-} from "viewer/model/actions/annotation_actions";
-import {
-  type MinCutAgglomerateAction,
-  type MinCutAgglomerateWithPositionAction,
-  type MinCutPartitionsAction,
   type ProofreadAtPositionAction,
-  type ProofreadMergeAction,
   resetMultiCutToolPartitionsAction,
   type ToggleSegmentInPartitionAction,
 } from "viewer/model/actions/proofread_actions";
-import {
-  dispatchEnsureHasNewestVersionAsync,
-  pushSaveQueueTransaction,
-  setPendingProofreadingOperationInfoAction,
-} from "viewer/model/actions/save_actions";
-import {
-  loadAdHocMeshAction,
-  loadPrecomputedMeshAction,
-} from "viewer/model/actions/segmentation_actions";
-import {
-  setMappingDataAction,
-  setMappingNameAction,
-  type UpdateUserSettingAction,
-} from "viewer/model/actions/settings_actions";
-import {
-  type CreateNodeAction,
-  type DeleteEdgeAction,
-  type DeleteNodeAction,
-  deleteEdgeAction,
-  type MergeTreesAction,
-  type SetNodePositionAction,
-  setTreeAgglomerateInfoIdAction,
-  setTreeNameAction,
-  setTreesAgglomerateInfoTracingIdAction,
+import type { UpdateUserSettingAction } from "viewer/model/actions/settings_actions";
+import type {
+  CreateNodeAction,
+  DeleteNodeAction,
+  SetNodePositionAction,
 } from "viewer/model/actions/skeletontracing_actions";
-import type { EnterAction, EscapeAction } from "viewer/model/actions/ui_actions";
+import type { EscapeAction } from "viewer/model/actions/ui_actions";
 import {
-  clickSegmentAction,
-  initializeEditableMappingAction,
-  mergeSegmentItemsAction,
-  removeSegmentAction,
   setActiveCellAction,
-  setHasEditableMappingAction,
   updateProofreadingMarkerPositionAction,
-  updateSegmentAction,
 } from "viewer/model/actions/volumetracing_actions";
 import type { Saga } from "viewer/model/sagas/effect_generators";
 import { select } from "viewer/model/sagas/effect_generators";
-import { getAgglomeratesForSegmentIds } from "viewer/model/sagas/volume/mapping_saga";
-import {
-  mergeAgglomerate,
-  splitAgglomerate,
-  type UpdateActionWithoutIsolationRequirement,
-} from "viewer/model/sagas/volume/update_actions";
-import { api, Model, Store } from "viewer/singletons";
-import type {
-  ActiveMappingInfo,
-  Mapping,
-  NumberLikeMap,
-  ProofreadingActionMappingInfo,
-  VolumeTracing,
-} from "viewer/store";
-import {
-  getAdditionalCoordinatesAsString,
-  getCurrentMag,
-} from "../../../accessors/flycam_accessor";
-import type { Action } from "../../../actions/actions";
-import type { Tree } from "../../../types/tree_types";
-import type { OperationContext } from "../../operation_context_saga";
+import { getAdditionalCoordinatesAsString } from "../../../accessors/flycam_accessor";
 import { ensureWkInitialized } from "../../ready_sagas";
+import { takeEveryInOperationContext, takeWithBatchActionSupport } from "../../saga_helpers";
+import { ensureHdf5MappingIsEnabled } from "./preparation_sagas";
 import {
-  spawnUntilCanceled,
-  takeEveryInOperationContext,
-  takeWithBatchActionSupport,
-} from "../../saga_helpers";
-import { subscribeToAnnotationMutex } from "../../saving/save_mutex_saga";
-import {
-  syncAgglomerateTreesAfterMergeAction,
-  syncAgglomerateTreesAfterSplitAction,
-} from "./agglomerate_tree_syncing_saga_helpers";
+  handleMinCutAgglomerate,
+  handleProofreadCutFromNeighbors,
+  handleProofreadMerge,
+  performPartitionedMinCut,
+} from "./proofread_action_handler_sagas";
+import { ensureSegmentItemAndMaybeLoadCoarseMesh } from "./segment_and_mesh_refresh_sagas";
+import { handleMergeViaTree, handleSplitViaTree } from "./tree_proofreading_sagas";
 
 function runSagaAndCatchSoftError<T>(saga: (...args: any[]) => Saga<T>) {
   return function* (...args: any[]) {
@@ -270,121 +178,6 @@ function* showToastIfSegmentOfOtherAgglomerateWasSelected(
   }
 }
 
-function proofreadCoarseMagIndex(): number {
-  // @ts-expect-error
-  return window.__proofreadCoarseResolutionIndex != null
-    ? // @ts-expect-error
-      window.__proofreadCoarseResolutionIndex
-    : 3;
-}
-
-function* syncWithBackend(ctx: OperationContext | undefined): Saga<void> {
-  yield* call([Model, Model.ensureSavedState], ctx);
-}
-
-function* pollNewestBackendVersion(ctx: OperationContext): Saga<void> {
-  yield* call(dispatchEnsureHasNewestVersionAsync, Store.dispatch, ctx);
-}
-
-function* subscribeToAnnotationMutexInLiveCollab(proofreadingSagaId: string) {
-  const isConcurrentCollab = yield* select(isConcurrentCollaborationMode);
-  if (isConcurrentCollab) {
-    return yield* call(subscribeToAnnotationMutex, proofreadingSagaId);
-  }
-  return null;
-}
-
-function* ensureSegmentItemAndMaybeLoadCoarseMesh(
-  layerName: string,
-  segmentId: number,
-  position: Vector3,
-  additionalCoordinates: AdditionalCoordinate[] | undefined,
-): Saga<void> {
-  yield* call(ensureSegmentItem, layerName, segmentId, position, additionalCoordinates);
-  const autoRenderMeshInProofreading = yield* select(
-    (state) => state.userConfiguration.autoRenderMeshInProofreading,
-  );
-  if (!autoRenderMeshInProofreading) {
-    return;
-  }
-  yield* call(loadCoarseMesh, layerName, segmentId, position, additionalCoordinates);
-}
-
-function* ensureSegmentItem(
-  layerName: string,
-  segmentId: number,
-  position: Vector3,
-  additionalCoordinates: AdditionalCoordinate[] | undefined,
-): Saga<void> {
-  yield* put(clickSegmentAction(segmentId, position, additionalCoordinates, layerName));
-}
-
-function* loadCoarseMesh(
-  layerName: string,
-  segmentId: number,
-  position: Vector3,
-  additionalCoordinates: AdditionalCoordinate[] | undefined,
-  opacity?: number,
-  isVisible?: boolean,
-): Saga<void> {
-  const dataset = yield* select((state) => state.dataset);
-  const layer = getLayerByName(dataset, layerName);
-
-  // Ensure that potential mesh files are already available. Otherwise, the following
-  // code would default to ad-hoc meshing.
-  yield* call(dispatchMaybeFetchMeshFilesAsync, Store.dispatch, layer, dataset, false);
-
-  const currentMeshFile = yield* select(
-    (state) => state.localSegmentationStateByLayer[layerName].currentMeshFile,
-  );
-
-  const meshInfo = yield* select((state) =>
-    getMeshInfoForSegment(state, additionalCoordinates || null, layerName, segmentId),
-  );
-
-  if (meshInfo != null) {
-    console.log(`Don't load mesh for segment ${segmentId} because it already exists.`);
-    return;
-  }
-
-  if (
-    currentMeshFile != null &&
-    currentMeshFile.formatVersion >= 3 &&
-    currentMeshFile.mappingName == null
-  ) {
-    // If a mesh file is active which was computed without a mapping, use that instead of computing
-    // meshes ad-hoc.
-    yield* put(
-      loadPrecomputedMeshAction(
-        segmentId,
-        position,
-        additionalCoordinates,
-        currentMeshFile.name,
-        opacity,
-        isVisible,
-        undefined,
-      ),
-    );
-  } else {
-    const mappingInfo = yield* select((state) =>
-      getMappingInfo(state.temporaryConfiguration.activeMappingByLayer, layerName),
-    );
-    const { mappingName, mappingType } = mappingInfo;
-
-    // Load the whole agglomerate mesh in a coarse mag for performance reasons
-    const preferredQuality = proofreadCoarseMagIndex();
-    yield* put(
-      loadAdHocMeshAction(segmentId, position, additionalCoordinates, {
-        mappingName,
-        mappingType,
-        preferredQuality,
-        opacity,
-        isVisible,
-      }),
-    );
-  }
-}
-
 function* checkForAgglomerateTreeModification(
   action: CreateNodeAction | DeleteNodeAction | SetNodePositionAction,
 ): Saga<void> {
@@ -404,47 +197,6 @@ function* checkForAgglomerateTreeModification(
       { timeout: 10000 },
     );
   }
-}
-
-function* pushPendingProofreadingOperationInfo(
-  volumeTracingId: string,
-  sourceInfo: ProofreadingActionMappingInfo,
-  targetInfo: ProofreadingActionMappingInfo | null = null,
-): Saga<void> {
-  // For proper post processing, it is necessary that sourceInfo & targetInfo (does make sense in splitting operations)
-  // have the correct agglomerate id values at the state on which this proofreading action
-  // is applied to. Due to first applying new update actions from the backend, the agglomerate id
-  // info might not be correct which is currently, stored in sourceInfo & targetInfo.
-  // Thus, in case the the saving first applies new missing update actions from the backend, we temporarily put
-  // the source- & targetInfo into the store, so that the save saga can update them after applying new backend actions
-  // and before the mapping changes of this proofreading action are applied.
-  // After saving the info should retrieved via the pop-variation of this function below.
-  yield* put(
-    setPendingProofreadingOperationInfoAction({
-      tracingId: volumeTracingId,
-      sourceInfo,
-      targetInfo,
-    }),
-  );
-}
-
-function* popPendingProofreadingOperationInfo(): Saga<
-  [ProofreadingActionMappingInfo, ProofreadingActionMappingInfo | null] | null
-> {
-  // See comment above.
-  // After the proofreading update actions were stored on the server, the updated source & targetInfo
-  // needs to retrieved and the PendingProofreadingOperationInfo should be cleared.
-  const pendingProofreadingOperationInfo = yield* select(
-    (state) => state.save.proofreadingPostProcessingInfo,
-  );
-  yield* put(setPendingProofreadingOperationInfoAction(null));
-  if (pendingProofreadingOperationInfo != null) {
-    return [
-      pendingProofreadingOperationInfo.sourceInfo,
-      pendingProofreadingOperationInfo.targetInfo,
-    ];
-  }
-  return null;
 }
 
 function* proofreadAtPosition(action: ProofreadAtPositionAction): Saga<void> {
@@ -473,894 +225,6 @@ function* proofreadAtPosition(action: ProofreadAtPositionAction): Saga<void> {
   );
 }
 
-export function* createEditableMapping(ctx?: OperationContext): Saga<string> {
-  /*
-   * Returns the name of the editable mapping. This is not identical to the
-   * name of the HDF5 mapping for which the editable mapping is about to be created.
-   */
-  // Get volume tracing again to make sure the version is up to date
-  const volumeTracing = yield* select((state) => getActiveSegmentationTracing(state));
-  if (!volumeTracing || !volumeTracing.mappingName) {
-    // This should never occur, because the proofreading tool is only available when a volume tracing layer is active.
-    throw new Error("No active segmentation tracing layer. Cannot create editable mapping.");
-  }
-
-  const volumeTracingId = volumeTracing.tracingId;
-  const layerName = volumeTracingId;
-  const baseMappingName = volumeTracing.mappingName;
-  yield* put(setMappingNameAction(layerName, volumeTracingId, "HDF5"));
-  yield* put(setHasEditableMappingAction(volumeTracingId));
-
-  // Ensure a saved state so that the mapping is locked and editable before doing the first proofreading operation.
-  // This needs to happen before initializeEditableMappingAction is dispatched, because the backend wouldn't be
-  // able to handle mapping requests that might be triggered right after initializing the mapping.
-  yield* call(syncWithBackend, ctx);
-
-  const editableMapping: ServerEditableMapping = {
-    baseMappingName: baseMappingName,
-    tracingId: volumeTracingId,
-    createdTimestamp: Date.now(),
-  };
-  yield* put(initializeEditableMappingAction(editableMapping));
-
-  yield* put(setTreesAgglomerateInfoTracingIdAction(volumeTracingId));
-  // Save, that the agglomerate info of all agglomerate trees was updated.
-  yield* call(syncWithBackend, ctx);
-
-  return volumeTracingId;
-}
-
-function* ensureHdf5MappingIsEnabled(layerName: string): Saga<boolean> {
-  const mappingInfo = yield* select((state) =>
-    getMappingInfo(state.temporaryConfiguration.activeMappingByLayer, layerName),
-  );
-  const { mappingName, mappingType, mappingStatus } = mappingInfo;
-  if (
-    mappingName == null ||
-    mappingType == null ||
-    mappingType === "JSON" ||
-    mappingStatus === MappingStatusEnum.DISABLED
-  ) {
-    Toast.error("An HDF5 mapping needs to be enabled to use the proofreading tool.");
-    return false;
-  }
-
-  return true;
-}
-
-function lookupAgglomerateId(
-  activeMapping: ActiveMappingInfo,
-  unmappedId: number,
-  fallback: number,
-): number {
-  if (activeMapping.mapping == null) return fallback;
-  return new NumberLikeMapWrapper(activeMapping.mapping).getAsNumber(unmappedId) ?? fallback;
-}
-
-// Shared setup for tree-based proofreading handlers. Returns null if the action should not proceed.
-// Note: the skeletontracing reducer already mutated the trees according to the received action.
-function* setupTreeProofreadingAction(
-  action: MergeTreesAction | DeleteEdgeAction | MinCutAgglomerateAction,
-  ctx: OperationContext,
-) {
-  const allowUpdate = yield* select(mayEditAnnotation);
-  if (!allowUpdate) return null;
-
-  // Reserve the mutex early to be able to synchronize the agglomerate trees with the newest backend state,
-  // in case of live collab being active.
-  // The reason is that in the meantime another user might have already manipulated the same agglomerate tree
-  // as done by the current action. Thus, to avoid inconsistent states between the mapping and the agglomerate trees,
-  // we first sync with the backend to have the newest agglomerate trees.
-  // Additionally, the skeletontracing reducer ignores direct proofreading agglomerate tree updates, in case live collab is active.
-  // We therefore explicitly replay the received actions after syncing with the backend, but setting this saga as the initiator.
-  // This should ensure, that the agglomerate tree updates done by the user are always in sync with the backend state.
-  const unsubscribeFromAnnotationMutex = yield* call(
-    subscribeToAnnotationMutexInLiveCollab,
-    "Proofreading via Trees",
-  );
-
-  let successfulReturn = false;
-  try {
-    const isLiveCollabActive = yield* select(isConcurrentCollaborationMode);
-    if (isLiveCollabActive) {
-      // If live collab is active and the user did a proofreading merge via edge creation / merge trees,
-      // the affected agglomerate tree(s) may not be in sync with the backend yet. So poll the latest updates.
-      // It is ensured that no new updates by other users are created during processing this proofreading
-      // interaction as we subscribed to the mutex above.
-      yield* call(pollNewestBackendVersion, ctx);
-    }
-    // Replay the current action with the proofreading saga as the initiator.
-    // The reducer ignores the current action as the initiator is marked as "USER".
-    const actionWithSagaAsInitiator = { ...action, initiator: "PROOFREADING" } as
-      | DeleteEdgeAction
-      | MergeTreesAction;
-    yield* put(actionWithSagaAsInitiator);
-
-    const { sourceNodeId, targetNodeId } = action;
-    const skeletonTracing = yield* select((state) => enforceSkeletonTracing(state.annotation));
-    const { trees } = skeletonTracing;
-    const sourceTree = findTreeByNodeId(trees, sourceNodeId);
-    const targetTree = findTreeByNodeId(trees, targetNodeId);
-
-    if (sourceTree == null || targetTree == null) {
-      return null;
-    }
-
-    const isModifyingOnlyAgglomerateTrees =
-      isAgglomerateTree(sourceTree) && isAgglomerateTree(targetTree);
-    const isModifyingAnyAgglomerateTrees =
-      isAgglomerateTree(sourceTree) || isAgglomerateTree(targetTree);
-    const isProofreadingToolActive = yield* select(
-      (state) => state.uiInformation.activeTool === AnnotationTool.PROOFREAD,
-    );
-
-    if (isProofreadingToolActive && !isModifyingOnlyAgglomerateTrees) {
-      Toast.warning(
-        "Only agglomerate trees can be modified using the proofreading tool to edit the active mapping.",
-        { timeout: 12000 },
-      );
-      return null;
-    } else if (!isProofreadingToolActive && isModifyingAnyAgglomerateTrees) {
-      Toast.warning(
-        "In order to edit the active mapping by deleting or adding edges of agglomerate trees, the proofreading tool needs to be active." +
-          " If you want to edit the active mapping, activate the proofreading tool and then redo the action.",
-        { timeout: 12000 },
-      );
-      return null;
-    }
-
-    if (!isProofreadingToolActive) {
-      return null;
-    }
-
-    const preparation = yield* call(prepareSplitOrMerge, true, ctx);
-    if (!preparation) {
-      return null;
-    }
-
-    // Use untransformedPosition because agglomerate trees should not have any transforms, anyway.
-    if (yield* select((state) => areGeometriesTransformed(state))) {
-      Toast.error(
-        "Proofreading is currently not supported when the skeleton layer is transformed.",
-      );
-      return null;
-    }
-
-    const sourceNodePosition = sourceTree.nodes.getOrThrow(sourceNodeId).untransformedPosition;
-    const targetNodePosition = targetTree.nodes.getOrThrow(targetNodeId).untransformedPosition;
-
-    const idInfos = yield* call(getAgglomerateInfos, preparation.getMappedAndUnmapped, [
-      sourceNodePosition,
-      targetNodePosition,
-    ]);
-    if (!idInfos) {
-      return null;
-    }
-
-    successfulReturn = true;
-    return {
-      unsubscribeFromAnnotationMutex,
-      preparation,
-      sourceTree,
-      targetTree,
-      sourceNodePosition,
-      targetNodePosition,
-      sourceNodeId,
-      targetNodeId,
-      idInfos,
-    };
-  } finally {
-    if (!successfulReturn && unsubscribeFromAnnotationMutex) {
-      yield* call(unsubscribeFromAnnotationMutex);
-    }
-  }
-}
-
-function* handleMergeViaTree(action: MergeTreesAction, ctx: OperationContext): Saga<void> {
-  // Ignore MergeTrees actions that were dispatched by the proofreading saga itself.
-  if (action.initiator === "PROOFREADING") return;
-
-  const setup = yield* call(setupTreeProofreadingAction, action, ctx);
-  if (!setup) return;
-  const {
-    unsubscribeFromAnnotationMutex,
-    preparation,
-    sourceNodePosition,
-    targetNodePosition,
-    sourceNodeId,
-    targetNodeId,
-    idInfos,
-  } = setup;
-
-  try {
-    const { getDataValue, volumeTracing } = preparation;
-    let { activeMapping } = preparation;
-    const { tracingId: volumeTracingId } = volumeTracing;
-
-    let [sourceInfo, targetInfo] = idInfos;
-    let sourceAgglomerateId = sourceInfo.agglomerateId;
-    let targetAgglomerateId = targetInfo.agglomerateId;
-
-    if (sourceAgglomerateId === targetAgglomerateId) {
-      Toast.error("Segments that should be merged need to be in different agglomerates.");
-      return;
-    }
-
-    /* Send the merge update action to the backend */
-    const items: UpdateActionWithoutIsolationRequirement[] = [
-      mergeAgglomerate(
-        sourceInfo.unmappedId,
-        targetInfo.unmappedId,
-        sourceAgglomerateId,
-        targetAgglomerateId,
-        volumeTracingId,
-      ),
-    ];
-    yield* call(
-      updateMappingWithMerge,
-      volumeTracingId,
-      activeMapping,
-      sourceAgglomerateId,
-      targetAgglomerateId,
-      false,
-    );
-
-    yield* call(pushPendingProofreadingOperationInfo, volumeTracingId, sourceInfo, targetInfo);
-    yield* put(pushSaveQueueTransaction(items));
-    const postProcessResult = yield* call(
-      syncAndUpdatePostProcessingInfo,
-      sourceInfo,
-      targetInfo,
-      ctx,
-    );
-    if (!postProcessResult) return;
-    ({ sourceInfo, targetInfo } = postProcessResult);
-
-    activeMapping = yield* select(
-      (store) => store.temporaryConfiguration.activeMappingByLayer[volumeTracing.tracingId],
-    );
-    sourceAgglomerateId = lookupAgglomerateId(
-      activeMapping,
-      sourceInfo.unmappedId,
-      sourceAgglomerateId,
-    );
-
-    const [newSourceAgglomerateId, newTargetAgglomerateId] = yield* all([
-      call(getDataValue, sourceNodePosition, activeMapping.mapping),
-      call(getDataValue, targetNodePosition, activeMapping.mapping),
-    ]);
-
-    let updatedSkeletonTracing = yield* select((state) => state.annotation.skeleton);
-    if (!updatedSkeletonTracing) {
-      Toast.error(
-        "Could not reload skeleton tracing to update the tree names and refresh the auxiliary meshes after proofreading action via agglomerate trees.",
-      );
-      return;
-    }
-
-    // Reload sourceTree & targetTree as applying newest actions from the backend might have modified the skeleton.
-    const updatedSourceTree = findTreeByNodeId(updatedSkeletonTracing.trees, sourceNodeId);
-    const updatedTargetTree = findTreeByNodeId(updatedSkeletonTracing.trees, targetNodeId);
-
-    if (updatedSourceTree == null || updatedTargetTree == null) {
-      Toast.error("Couldn't find trees for nodes. Details are logged to the console.");
-      console.error("Couldn't find trees for nodes. Details are logged to the console.", {
-        sourceNodeId,
-        updatedSourceTree,
-        targetNodeId,
-        updatedTargetTree,
-      });
-      return;
-    }
-
-    /* Rename agglomerate tree according to its new id and mapping name */
-    yield* put(
-      setTreeNameAction(
-        getTreeNameForAgglomerateTree(newSourceAgglomerateId, volumeTracing.mappingName),
-        updatedSourceTree.treeId,
-      ),
-    );
-    yield* put(setTreeAgglomerateInfoIdAction(newSourceAgglomerateId, updatedSourceTree.treeId));
-
-    // Also merge the segment items.
-    yield* put(
-      mergeSegmentItemsAction(
-        sourceInfo.agglomerateId,
-        targetInfo.agglomerateId,
-        sourceInfo.unmappedId,
-        targetInfo.unmappedId,
-        volumeTracingId,
-      ),
-    );
-
-    /* Ensure segment items exist for affected segments and reload affected meshes */
-    const refreshInfos = [
-      {
-        oldAgglomerateId: sourceInfo.agglomerateId,
-        newAgglomerateId: newSourceAgglomerateId,
-        nodePosition: sourceNodePosition,
-      },
-      {
-        oldAgglomerateId: targetInfo.agglomerateId,
-        newAgglomerateId: newTargetAgglomerateId,
-        nodePosition: targetNodePosition,
-      },
-    ];
-    yield* call(refreshAffectedSegmentItems, volumeTracingId, refreshInfos);
-    // Now that the segment items are up-to-date we can sync with the back-end and release the mutex.
-    yield* call(syncWithBackend, ctx);
-
-    // Refreshing the meshes might take a while and won't block the saga here.
-    yield* spawnUntilCanceled(maybeRefreshAffectedMeshes, volumeTracingId, refreshInfos);
-  } finally {
-    if (unsubscribeFromAnnotationMutex) {
-      yield* call(unsubscribeFromAnnotationMutex);
-    }
-  }
-}
-
-function* handleSplitViaTree(
-  action: DeleteEdgeAction | MinCutAgglomerateAction,
-  ctx: OperationContext,
-): Saga<void> {
-  // Ignore DeleteEdge actions that were dispatched by the proofreading saga itself.
-  if (action.type === "DELETE_EDGE" && action.initiator === "PROOFREADING") return;
-
-  const setup = yield* call(setupTreeProofreadingAction, action, ctx);
-  if (!setup) return;
-
-  const {
-    unsubscribeFromAnnotationMutex,
-    preparation,
-    sourceTree,
-    sourceNodePosition,
-    targetNodePosition,
-    sourceNodeId,
-    targetNodeId,
-    idInfos,
-  } = setup;
-  try {
-    const { agglomerateFileMag, getDataValue, volumeTracing, annotationVersion } = preparation;
-    let { activeMapping } = preparation;
-    const { tracingId: volumeTracingId } = volumeTracing;
-
-    let [sourceInfo, targetInfo] = idInfos;
-    let sourceAgglomerateId = sourceInfo.agglomerateId;
-    let targetAgglomerateId = targetInfo.agglomerateId;
-
-    /* Send the split update action to the backend */
-    const items: UpdateActionWithoutIsolationRequirement[] = [];
-    if (action.type === "DELETE_EDGE") {
-      if (sourceAgglomerateId !== targetAgglomerateId) {
-        Toast.error("Segments that should be split need to be in the same agglomerate.");
-        return;
-      }
-      items.push(
-        splitAgglomerate(
-          sourceInfo.unmappedId,
-          targetInfo.unmappedId,
-          sourceAgglomerateId,
-          volumeTracingId,
-        ),
-      );
-    } else {
-      const [hasErrored] = yield* call(
-        performMinCut,
-        sourceAgglomerateId,
-        targetAgglomerateId,
-        [sourceInfo.unmappedId],
-        [targetInfo.unmappedId],
-        agglomerateFileMag,
-        volumeTracingId,
-        sourceTree,
-        annotationVersion,
-        items,
-      );
-      if (hasErrored) {
-        return;
-      }
-    }
-
-    if (items.length === 0) {
-      return;
-    }
-
-    yield* call(pushPendingProofreadingOperationInfo, volumeTracingId, sourceInfo, targetInfo);
-    yield* put(pushSaveQueueTransaction(items));
-    const postProcessResult = yield* call(
-      syncAndUpdatePostProcessingInfo,
-      sourceInfo,
-      targetInfo,
-      ctx,
-    );
-    if (!postProcessResult) return;
-    ({ sourceInfo, targetInfo } = postProcessResult);
-
-    activeMapping = yield* select(
-      (store) => store.temporaryConfiguration.activeMappingByLayer[volumeTracing.tracingId],
-    );
-    sourceAgglomerateId = lookupAgglomerateId(
-      activeMapping,
-      sourceInfo.unmappedId,
-      sourceAgglomerateId,
-    );
-    targetAgglomerateId = lookupAgglomerateId(
-      activeMapping,
-      targetInfo.unmappedId,
-      targetAgglomerateId,
-    );
-
-    if (sourceAgglomerateId !== targetAgglomerateId) {
-      Toast.error(
-        "The selected positions are not part of the same agglomerate and cannot be split.",
-      );
-      return;
-    }
-
-    // Because we ensured a saved state a few lines above, we can now split the mapping locally
-    // as this still requires some communication with the back-end.
-    const currentAnnotationVersion = yield* select((state) => state.annotation.version);
-    const splitMappingInfo = yield* splitAgglomerateInMapping(
-      activeMapping,
-      sourceAgglomerateId,
-      [sourceInfo.unmappedId],
-      volumeTracingId,
-      currentAnnotationVersion,
-      // No need to refresh the agglomerate meshes as this split is triggered by a previous
-      // delete edge action, already correctly updating the agglomerate tree.
-      false,
-    );
-    if (splitMappingInfo == null) {
-      console.error("Failed to split mapping in tree based proofreading action. Aborting...");
-      return;
-    }
-
-    const { mappingWithSplitApplied } = splitMappingInfo;
-    yield* put(
-      setMappingDataAction(
-        volumeTracingId,
-        mappingWithSplitApplied,
-        // As these split actions were already sent to the server, mappingWithSplitApplied is stored on the server already.
-        true,
-      ),
-    );
-
-    const [newSourceAgglomerateId, newTargetAgglomerateId] = yield* all([
-      call(getDataValue, sourceNodePosition, mappingWithSplitApplied),
-      call(getDataValue, targetNodePosition, mappingWithSplitApplied),
-    ]);
-
-    if (newSourceAgglomerateId === newTargetAgglomerateId) {
-      // The split was unsuccessful.
-      Toast.warning("The split operation was unsuccessful. Please retry.");
-      yield* call(syncWithBackend, ctx);
-      return;
-    }
-
-    let updatedSkeletonTracing = yield* select((state) => state.annotation.skeleton);
-    if (!updatedSkeletonTracing) {
-      Toast.error(
-        "Could not reload skeleton tracing to update the tree names and refresh the auxiliary meshes after proofreading action via agglomerate trees.",
-      );
-      return;
-    }
-
-    // Reload trees as applying newest actions from the backend might have modified the skeleton.
-    const updatedSourceTree = findTreeByNodeId(updatedSkeletonTracing.trees, sourceNodeId);
-    const updatedTargetTree = findTreeByNodeId(updatedSkeletonTracing.trees, targetNodeId);
-
-    if (updatedSourceTree == null || updatedTargetTree == null) {
-      Toast.error("Couldn't find trees for nodes. Details are logged to the console.");
-      console.error("Couldn't find trees for nodes. Details are logged to the console.", {
-        sourceNodeId,
-        updatedSourceTree,
-        targetNodeId,
-        updatedTargetTree,
-      });
-      return;
-    }
-
-    /* Rename agglomerate trees according to their new ids and mapping name */
-    yield* put(
-      setTreeNameAction(
-        getTreeNameForAgglomerateTree(newSourceAgglomerateId, volumeTracing.mappingName),
-        updatedSourceTree.treeId,
-      ),
-    );
-    yield* put(setTreeAgglomerateInfoIdAction(newSourceAgglomerateId, updatedSourceTree.treeId));
-    // A split happened: rename the target tree and create a segment for it.
-    yield* put(
-      setTreeNameAction(
-        getTreeNameForAgglomerateTree(newTargetAgglomerateId, volumeTracing.mappingName),
-        updatedTargetTree.treeId,
-      ),
-    );
-    yield* put(setTreeAgglomerateInfoIdAction(newTargetAgglomerateId, updatedTargetTree.treeId));
-    const newSegmentName =
-      (yield* select(
-        (state) =>
-          getVolumeTracingById(state.annotation, volumeTracingId).segments.getNullable(
-            sourceInfo.agglomerateId,
-          )?.name,
-      )) ?? `Agglomerate ${newTargetAgglomerateId}`;
-    yield* put(
-      updateSegmentAction(
-        Number(newTargetAgglomerateId),
-        { name: newSegmentName },
-        volumeTracingId,
-      ),
-    );
-
-    /* Ensure segment items exist for affected segments and reload affected meshes */
-    const refreshInfos = [
-      {
-        oldAgglomerateId: sourceInfo.agglomerateId,
-        newAgglomerateId: newSourceAgglomerateId,
-        nodePosition: sourceNodePosition,
-      },
-      {
-        oldAgglomerateId: targetInfo.agglomerateId,
-        newAgglomerateId: newTargetAgglomerateId,
-        nodePosition: targetNodePosition,
-      },
-    ];
-    yield* call(refreshAffectedSegmentItems, volumeTracingId, refreshInfos);
-    // Now that the segment items are up-to-date we can sync with the back-end and release the mutex.
-    yield* call(syncWithBackend, ctx);
-
-    // Refreshing the meshes might take a while and won't block the saga here.
-    yield* spawnUntilCanceled(maybeRefreshAffectedMeshes, volumeTracingId, refreshInfos);
-  } finally {
-    if (unsubscribeFromAnnotationMutex) {
-      yield* call(unsubscribeFromAnnotationMutex);
-    }
-  }
-}
-
-// Returns a tuple of whether the min cut failed and if successful a list of edges removed by the min cut.
-function* performMinCut(
-  sourceAgglomerateId: number,
-  targetAgglomerateId: number,
-  sourceSegmentIds: number[],
-  targetSegmentIds: number[],
-  agglomerateFileMag: Vector3,
-  volumeTracingId: string,
-  sourceTree: Tree | null,
-  version: number,
-  items: UpdateActionWithoutIsolationRequirement[],
-): Saga<[boolean, MinCutTargetEdge[]]> {
-  if (sourceAgglomerateId !== targetAgglomerateId) {
-    Toast.error(
-      `Segments need to be in the same agglomerate to perform a min-cut splitting operation. Agglomerate ids are ${sourceAgglomerateId} and ${targetAgglomerateId}.`,
-    );
-    return [true, []];
-  }
-
-  const tracingStoreUrl = yield* select((state) => state.annotation.tracingStore.url);
-  const segmentsInfo = {
-    partition1: sourceSegmentIds,
-    partition2: targetSegmentIds,
-    mag: agglomerateFileMag,
-    agglomerateId: sourceAgglomerateId,
-    editableMappingId: volumeTracingId,
-  };
-
-  let edgesToRemove: MinCutTargetEdge[] = [];
-  try {
-    edgesToRemove = yield* call(
-      getEdgesForAgglomerateMinCut,
-      tracingStoreUrl,
-      volumeTracingId,
-      version,
-      segmentsInfo,
-    );
-  } catch (exception) {
-    console.error(exception);
-    Toast.error("Could not determine which edges to delete for cut. Please try again.");
-    return [true, []];
-  }
-
-  // Use untransformedPosition below because agglomerate trees should not have
-  // any transforms, anyway.
-  if (yield* select((state) => areGeometriesTransformed(state))) {
-    Toast.error("Proofreading is currently not supported when the skeleton layer is transformed.");
-    return [true, []];
-  }
-
-  for (const edge of edgesToRemove) {
-    if (sourceTree) {
-      const result = getDeleteEdgeActionForEdgePositions(sourceTree, edge);
-      if (result == null) {
-        return [true, []];
-      }
-      const { firstNodeId, secondNodeId } = result;
-      yield* put(deleteEdgeAction(firstNodeId, secondNodeId, Date.now(), "PROOFREADING"));
-    }
-
-    console.log(
-      "Splitting agglomerate",
-      sourceAgglomerateId,
-      "with segment ids",
-      edge.segmentId1,
-      "and",
-      edge.segmentId2,
-    );
-    items.push(
-      splitAgglomerate(edge.segmentId1, edge.segmentId2, sourceAgglomerateId, volumeTracingId),
-    );
-  }
-
-  return [false, edgesToRemove];
-}
-
-function* performPartitionedMinCut(
-  action: MinCutPartitionsAction | EnterAction,
-  ctx: OperationContext,
-): Saga<void> {
-  const isMultiSplitActive = yield* select((state) => state.userConfiguration.isMultiSplitActive);
-  const isFromEditingEvent = action.type === "ENTER" && isEditableEventTarget(action.event.target);
-  const activeTool = yield* select((state) => state.uiInformation.activeTool);
-  if (!isMultiSplitActive || activeTool !== AnnotationTool.PROOFREAD || isFromEditingEvent) {
-    return;
-  }
-
-  const preparation = yield* call(prepareSplitOrMerge, false, ctx);
-  if (!preparation) {
-    return;
-  }
-  const partitions = yield* select(
-    (state) =>
-      state.localSegmentationStateByLayer[preparation.volumeTracing.tracingId].minCutPartitions,
-  );
-  let agglomerateId = partitions.agglomerateId;
-  if (partitions[1].length <= 0 || partitions[2].length <= 0) {
-    console.error(messages["proofreading.multi_cut.empty_partition"]);
-    Toast.error(messages["proofreading.multi_cut.empty_partition"]);
-    return;
-  }
-  if (agglomerateId == null) {
-    console.error(messages["proofreading.multi_cut.no_valid_agglomerate"]);
-    Toast.error(messages["proofreading.multi_cut.no_valid_agglomerate"]);
-    return;
-  }
-  const volumeTracingId = preparation.volumeTracing.tracingId;
-  const { agglomerateFileMag, annotationVersion } = preparation;
-
-  const items: UpdateActionWithoutIsolationRequirement[] = [];
-
-  const [hasErrored, edgesToRemove] = yield* call(
-    performMinCut,
-    agglomerateId,
-    agglomerateId,
-    partitions[1],
-    partitions[2],
-    agglomerateFileMag,
-    volumeTracingId,
-    null,
-    annotationVersion,
-    items,
-  );
-  if (hasErrored || edgesToRemove.length === 0) {
-    console.error(messages["proofreading.multi_cut.split_failed"]);
-    Toast.error(messages["proofreading.multi_cut.split_failed"]);
-    return;
-  }
-
-  // Only one info object is needed as the min cut is performed on only one
-  const dummySourceInfo = { agglomerateId, unmappedId: partitions[1][0] };
-
-  yield* call(pushPendingProofreadingOperationInfo, volumeTracingId, dummySourceInfo);
-
-  yield* put(pushSaveQueueTransaction(items));
-
-  const unsubscribeFromAnnotationMutex = yield* call(
-    subscribeToAnnotationMutexInLiveCollab,
-    "Proofreading Partitioned Min-Cut",
-  );
-  try {
-    const postProcessResult = yield* call(
-      syncAndUpdatePostProcessingInfo,
-      dummySourceInfo,
-      dummySourceInfo,
-      ctx,
-    );
-    if (!postProcessResult) return;
-    const agglomerateIdBeforeSplit = postProcessResult.sourceInfo.agglomerateId;
-
-    yield* put(resetMultiCutToolPartitionsAction());
-
-    const activeMapping = yield* select(
-      (store) => store.temporaryConfiguration.activeMappingByLayer[volumeTracingId],
-    );
-
-    // The agglomerateId of the split agglomerate might have changed due to syncing with the server caused by Model.ensureSavedState.
-    // Thus we reload the agglomerateId via simply looking it up via the first segment of partition 1.
-    agglomerateId = lookupAgglomerateId(activeMapping, partitions[1][0], agglomerateId);
-
-    const unmappedSegmentsOfPartitions = new Set([...partitions[1], ...partitions[2]]);
-    // Make sure the reloaded partial mapping has mapping info about the partitions and first removed edge. The first removed edge is used for reloading the meshes.
-    // The unmapped segments of this edge might not be present in the partial mapping of the frontend as splitting can be done via mesh interactions.
-    // There is no guarantee that for all mesh parts the mapping is locally stored.
-    // So we put them all into segmentsInvolvedInSplit.
-    // All these segments belonged to agglomerateId before the (min-cut) split.
-    const segmentsInvolvedInSplit = unmappedSegmentsOfPartitions.union(
-      new Set([edgesToRemove[0].segmentId1, edgesToRemove[0].segmentId2]),
-    );
-
-    // Now that the changes are saved, we can split the mapping locally (because it requires
-    // communication with the backend).
-    const currentVersion = Store.getState().annotation.version;
-
-    const autoUpdateAgglomerateTrees = true;
-    const splitMappingInfo = yield* splitAgglomerateInMapping(
-      activeMapping,
-      agglomerateId,
-      segmentsInvolvedInSplit,
-      volumeTracingId,
-      currentVersion,
-      autoUpdateAgglomerateTrees,
-      // Min-cut reads the partition/edge segments back from the mapping, so they must be added.
-      true,
-    );
-    if (splitMappingInfo == null) {
-      console.error("Failed to split mapping in partitioned min cut. Aborting...");
-      return;
-    }
-    const { mappingWithSplitApplied } = splitMappingInfo;
-
-    yield* put(
-      setMappingDataAction(
-        volumeTracingId,
-        mappingWithSplitApplied,
-        // As these split actions were already sent to the server, mappingWithSplitApplied is stored on the server already.
-        true,
-      ),
-    );
-
-    /* Reload meshes */
-    const newAgglomerateIdFromPartition1 = yield* call(
-      preparation.mapSegmentId,
-      partitions[1][0],
-      mappingWithSplitApplied,
-    );
-    const newAgglomerateIdFromPartition2 = yield* call(
-      preparation.mapSegmentId,
-      partitions[2][0],
-      mappingWithSplitApplied,
-    );
-
-    // Get positions of new meshes from first split edge information.
-    const firstEdgeFirstSegmentNewAgglomerate = yield* call(
-      preparation.mapSegmentId,
-      edgesToRemove[0].segmentId1,
-      mappingWithSplitApplied,
-    );
-    const meshLoadingPositionForPartition1 =
-      firstEdgeFirstSegmentNewAgglomerate === newAgglomerateIdFromPartition1
-        ? edgesToRemove[0].position1
-        : edgesToRemove[0].position2;
-    const meshLoadingPositionForPartition2 =
-      firstEdgeFirstSegmentNewAgglomerate === newAgglomerateIdFromPartition2
-        ? edgesToRemove[0].position1
-        : edgesToRemove[0].position2;
-
-    /* Ensure segment items exist for affected segments and reload affected meshes */
-    const refreshInfos = [
-      {
-        oldAgglomerateId: agglomerateIdBeforeSplit,
-        newAgglomerateId: newAgglomerateIdFromPartition1,
-        nodePosition: meshLoadingPositionForPartition1,
-      },
-      {
-        oldAgglomerateId: agglomerateIdBeforeSplit,
-        newAgglomerateId: newAgglomerateIdFromPartition2,
-        nodePosition: meshLoadingPositionForPartition2,
-      },
-    ];
-    yield* call(refreshAffectedSegmentItems, volumeTracingId, refreshInfos);
-
-    // Now that the segment items are up-to-date we can sync with the back-end
-    // and release the mutex.
-    yield* call(syncWithBackend, ctx);
-
-    // Refreshing the meshes might take a while and won't block the saga
-    // here.
-    yield* spawnUntilCanceled(maybeRefreshAffectedMeshes, volumeTracingId, refreshInfos);
-  } finally {
-    if (unsubscribeFromAnnotationMutex) {
-      yield* call(unsubscribeFromAnnotationMutex);
-    }
-  }
-}
-
-function* performCutFromNeighbors(
-  agglomerateId: number,
-  segmentId: number,
-  segmentPosition: Vector3 | null,
-  agglomerateFileMag: Vector3,
-  volumeTracingId: string,
-  annotationVersion: number,
-  sourceTree: Tree | null | undefined,
-  items: UpdateActionWithoutIsolationRequirement[],
-): Saga<
-  { didCancel: false; neighborInfo: NeighborInfo } | { didCancel: true; neighborInfo?: null }
-> {
-  const tracingStoreUrl = yield* select((state) => state.annotation.tracingStore.url);
-  const segmentsInfo = {
-    segmentId,
-    mag: agglomerateFileMag,
-    agglomerateId,
-    editableMappingId: volumeTracingId,
-  };
-
-  let neighborInfo;
-  try {
-    neighborInfo = yield* call(
-      getNeighborsForAgglomerateNode,
-      tracingStoreUrl,
-      volumeTracingId,
-      annotationVersion,
-      segmentsInfo,
-    );
-  } catch (exception) {
-    console.error(exception);
-    Toast.error("Could not load neighbors of agglomerate node. Please try again.");
-    return { didCancel: true };
-  }
-
-  const edgesToRemove: Array<
-    | {
-        position1: Vector3;
-        position2: Vector3;
-        segmentId1: number;
-        segmentId2: number;
-      }
-    | {
-        position1: null;
-        position2: Vector3;
-        segmentId1: number;
-        segmentId2: number;
-      }
-  > = neighborInfo.neighbors.map(
-    (neighbor) =>
-      ({
-        position1: segmentPosition,
-        position2: neighbor.position,
-        segmentId1: segmentId,
-        segmentId2: neighbor.segmentId,
-      }) as const,
-  );
-
-  if (edgesToRemove.length === 0) {
-    Toast.info("No neighbors found.");
-    return { didCancel: true };
-  }
-
-  for (const edge of edgesToRemove) {
-    if (sourceTree) {
-      if (edge.position1 == null) {
-        // Satisfy TypeScript. Should not happen because segmentPosition should not be null
-        // when a sourceTree was passed.
-        Toast.warning("Could not perform cut from neighbors. See console for more details.");
-        console.warn(
-          "segmentPosition is not available even though a tree was passed to performCutFromNeighbors.",
-        );
-        return { didCancel: true };
-      }
-      const result = getDeleteEdgeActionForEdgePositions(sourceTree, edge);
-      if (result == null) {
-        return { didCancel: true };
-      }
-      const { firstNodeId, secondNodeId } = result;
-      yield* put(deleteEdgeAction(firstNodeId, secondNodeId, Date.now(), "PROOFREADING"));
-    }
-
-    items.push(splitAgglomerate(edge.segmentId1, edge.segmentId2, agglomerateId, volumeTracingId));
-  }
-
-  return { didCancel: false, neighborInfo };
-}
-
 function* clearProofreadingByproducts() {
   const volumeTracingLayer = yield* select((state) => getActiveSegmentationTracingLayer(state));
   if (volumeTracingLayer == null || volumeTracingLayer.tracingId == null) return;
@@ -1379,1316 +243,4 @@ function* clearProofreadingByproducts() {
   for (const action of meshRemoveActions) {
     yield* put(action);
   }
-}
-
-const MISSING_INFORMATION_WARNING =
-  "Please use either the data viewports OR the 3D viewport (but not both) for selecting the partners of a proofreading operation.";
-
-function* syncAndUpdatePostProcessingInfo<
-  T1 extends IdInfo | IdInfoOpt | IdInfoWithoutPosition,
-  T2 extends IdInfo | IdInfoOpt | IdInfoWithoutPosition,
->(
-  sourceInfo: T1,
-  targetInfo: T2,
-  ctx: OperationContext,
-): Saga<{ sourceInfo: T1; targetInfo: T2 } | null> {
-  yield* call(syncWithBackend, ctx);
-  const proofreadingPostProcessingInfo = yield* call(popPendingProofreadingOperationInfo);
-  if (!proofreadingPostProcessingInfo) {
-    Toast.error(messages["proofreading.post_processing_info_not_found"]);
-    return null;
-  }
-  const updatedSourceInfo = {
-    ...sourceInfo,
-    agglomerateId: proofreadingPostProcessingInfo[0].agglomerateId,
-  };
-  const updatedTargetInfo = proofreadingPostProcessingInfo[1]
-    ? { ...targetInfo, agglomerateId: proofreadingPostProcessingInfo[1].agglomerateId }
-    : targetInfo;
-  return { sourceInfo: updatedSourceInfo, targetInfo: updatedTargetInfo };
-}
-
-function* refreshProofreadingSegmentsAndMeshes(
-  volumeTracingId: string,
-  sourceInfo: IdInfo,
-  targetInfo: IdInfoOpt,
-  sourceAgglomerateId: number,
-  targetAgglomerateId: number,
-  ctx: OperationContext,
-): Saga<void> {
-  /* Ensure segment items exist for affected segments and reload affected meshes */
-  const refreshInfos = [
-    {
-      oldAgglomerateId: sourceInfo.agglomerateId,
-      newAgglomerateId: sourceAgglomerateId,
-      nodePosition: sourceInfo.position,
-    },
-    {
-      oldAgglomerateId: targetInfo.agglomerateId,
-      newAgglomerateId: targetAgglomerateId,
-      nodePosition:
-        // targetInfo.position can only be undefined in case of
-        // a merge (see idInfos.type). In that case,
-        // this element was merged into another element.
-        // Therefore, sourceInfo.position is a valid replacement.
-        targetInfo.position ?? sourceInfo.position,
-    },
-  ];
-  yield* call(refreshAffectedSegmentItems, volumeTracingId, refreshInfos);
-  yield* call(syncWithBackend, ctx);
-
-  // Refreshing the meshes might take a while and won't block the saga here.
-  yield* spawnUntilCanceled(maybeRefreshAffectedMeshes, volumeTracingId, refreshInfos);
-}
-
-function* handleProofreadMerge(action: ProofreadMergeAction, ctx: OperationContext) {
-  const allowUpdate = yield* select((state) => mayEditAnnotation(state));
-  if (!allowUpdate) return;
-
-  const preparation = yield* call(prepareSplitOrMerge, false, ctx);
-  if (!preparation) return;
-
-  let { volumeTracing, activeMapping } = preparation;
-  const { tracingId: volumeTracingId } = volumeTracing;
-  const idInfos = yield* call(gatherInfoForOperation, action, preparation);
-
-  if (idInfos == null) {
-    console.warn("[Proofreading] Could not gather id infos.");
-    return;
-  }
-  let [sourceInfo, targetInfo] = idInfos.infos;
-  let sourceAgglomerateId = sourceInfo.agglomerateId;
-  let targetAgglomerateId = targetInfo.agglomerateId;
-
-  if (sourceAgglomerateId === targetAgglomerateId) {
-    Toast.error(
-      `Segments that should be merged need to be in different agglomerates. Both segments belong to agglomerate id=${sourceAgglomerateId}`,
-    );
-    return;
-  }
-
-  /* Send the merge update action to the backend (by pushing to the save queue and saving immediately) */
-  const updateActions: UpdateActionWithoutIsolationRequirement[] = [
-    mergeAgglomerate(
-      sourceInfo.unmappedId,
-      targetInfo.unmappedId,
-      sourceAgglomerateId,
-      targetAgglomerateId,
-      volumeTracingId,
-    ),
-  ];
-
-  console.log(
-    "Merging agglomerate",
-    sourceAgglomerateId,
-    "with",
-    targetAgglomerateId,
-    "and segment ids",
-    sourceInfo.unmappedId,
-    targetInfo.unmappedId,
-  );
-  yield* call(
-    updateMappingWithMerge,
-    volumeTracingId,
-    activeMapping,
-    sourceAgglomerateId,
-    targetAgglomerateId,
-    false,
-  );
-
-  yield* call(pushPendingProofreadingOperationInfo, volumeTracingId, sourceInfo, targetInfo);
-  yield* put(pushSaveQueueTransaction(updateActions));
-  const unsubscribeFromAnnotationMutex = yield* call(
-    subscribeToAnnotationMutexInLiveCollab,
-    "Proofreading Merge",
-  );
-  try {
-    // Remove the segment that doesn't exist anymore.
-    yield* put(
-      mergeSegmentItemsAction(
-        sourceAgglomerateId,
-        targetAgglomerateId,
-        sourceInfo.unmappedId,
-        targetInfo.unmappedId,
-        volumeTracingId,
-      ),
-    );
-
-    const postProcessResult = yield* call(() =>
-      syncAndUpdatePostProcessingInfo(sourceInfo, targetInfo, ctx),
-    );
-    if (!postProcessResult) return;
-    ({ sourceInfo, targetInfo } = postProcessResult);
-
-    // After saving and thus syncing with the server the mapping might have updated due to missing proofreading actions for other users.
-    // Thus, the sourceAgglomerateId and targetAgglomerateId might be outdated. Therefore, we reload them.
-    const newInfo = yield* call(
-      reloadMappingAndAggloIds,
-      volumeTracing.tracingId,
-      sourceInfo.unmappedId,
-      targetInfo.unmappedId,
-    );
-    if (!newInfo) {
-      Toast.error(
-        `Could not reload the agglomerate information for proofreading operation segments.`,
-      );
-      return;
-    }
-    activeMapping = newInfo.activeMapping;
-    sourceAgglomerateId = newInfo.sourceAgglomerateId;
-    targetAgglomerateId = newInfo.targetAgglomerateId;
-
-    // Reload the agglomerate trees affected by the merge if they are loaded.
-    yield* call(
-      syncAgglomerateTreesAfterMergeAction,
-      sourceInfo.agglomerateId,
-      targetInfo.agglomerateId,
-      sourceAgglomerateId,
-      volumeTracingId,
-    );
-
-    yield* call(
-      refreshProofreadingSegmentsAndMeshes,
-      volumeTracingId,
-      sourceInfo,
-      targetInfo,
-      sourceAgglomerateId,
-      targetAgglomerateId,
-      ctx,
-    );
-  } finally {
-    if (unsubscribeFromAnnotationMutex) {
-      yield* call(unsubscribeFromAnnotationMutex);
-    }
-  }
-}
-
-function* handleMinCutAgglomerate(
-  action: MinCutAgglomerateWithPositionAction,
-  ctx: OperationContext,
-) {
-  const allowUpdate = yield* select((state) => mayEditAnnotation(state));
-  if (!allowUpdate) return;
-
-  const preparation = yield* call(prepareSplitOrMerge, false, ctx);
-  if (!preparation) return;
-
-  let { agglomerateFileMag, volumeTracing, activeMapping, annotationVersion } = preparation;
-  const { tracingId: volumeTracingId } = volumeTracing;
-  const idInfos = yield* call(gatherInfoForOperation, action, preparation);
-
-  if (idInfos == null) {
-    console.warn("[Proofreading] Could not gather id infos.");
-    return;
-  }
-  let [sourceInfo, targetInfo] = idInfos.infos;
-  let sourceAgglomerateId = sourceInfo.agglomerateId;
-  let targetAgglomerateId = targetInfo.agglomerateId;
-
-  if (sourceInfo.unmappedId === targetInfo.unmappedId) {
-    Toast.error(
-      "The selected positions are both part of the same base segment and cannot be split. Please select another position or use the nodes of the agglomerate tree to perform the split.",
-    );
-    return;
-  }
-
-  /* Send the split update action to the backend (by pushing to the save queue and saving immediately) */
-  const updateActions: UpdateActionWithoutIsolationRequirement[] = [];
-  const [hasErrored] = yield* call(
-    performMinCut,
-    sourceAgglomerateId,
-    targetAgglomerateId,
-    [sourceInfo.unmappedId],
-    [targetInfo.unmappedId],
-    agglomerateFileMag,
-    volumeTracingId,
-    null,
-    annotationVersion,
-    updateActions,
-  );
-  if (hasErrored) {
-    console.error(messages["proofreading.multi_cut.split_failed"]);
-    Toast.error(messages["proofreading.multi_cut.split_failed"]);
-    return;
-  }
-
-  if (updateActions.length === 0) {
-    return;
-  }
-  yield* call(pushPendingProofreadingOperationInfo, volumeTracingId, sourceInfo, targetInfo);
-  yield* put(pushSaveQueueTransaction(updateActions));
-  const unsubscribeFromAnnotationMutex = yield* call(
-    subscribeToAnnotationMutexInLiveCollab,
-    "Proofreading Min-Cut",
-  );
-  try {
-    const postProcessResult = yield* call(() =>
-      syncAndUpdatePostProcessingInfo(sourceInfo, targetInfo, ctx),
-    );
-    if (!postProcessResult) return;
-    ({ sourceInfo, targetInfo } = postProcessResult);
-
-    if (sourceAgglomerateId !== targetAgglomerateId) {
-      const othersMayEdit = yield* select((state) =>
-        isAnnotationEditableByNonOwners(state.annotation),
-      );
-      const additionalErrorExplanation = othersMayEdit
-        ? " Maybe another user already split the agglomerate in the meantime."
-        : "";
-      Toast.error(
-        `The selected positions are not part of the same agglomerate and cannot be split.${additionalErrorExplanation}`,
-      );
-      return;
-    }
-
-    // After saving and thus syncing with the server the mapping might have updated due to missing proofreading actions for other users.
-    // Thus, the sourceAgglomerateId and targetAgglomerateId might be outdated. Therefore, we reload them.
-    annotationVersion = yield* select((state) => state.annotation.version);
-
-    // Get latest agglomerate id of the agglomerate that was split by the current operation in case it changed due to rebasing or so.
-    const latestInfo = yield* call(
-      reloadMappingAndAggloIds,
-      volumeTracing.tracingId,
-      sourceInfo.unmappedId,
-      targetInfo.unmappedId,
-    );
-    if (!latestInfo) {
-      Toast.error(
-        `Could not reload the agglomerate information for proofreading split segments operation.`,
-      );
-      return;
-    }
-    activeMapping = latestInfo.activeMapping;
-    const latestSourceAgglomerateId = latestInfo.sourceAgglomerateId;
-
-    console.log("start updating the mapping after a min-cut");
-    // Now that the changes are saved, we can split the local mapping (because it requires
-    // communication with the back-end).
-    const autoUpdateAgglomerateTrees = true;
-    const splitMappingInfo = yield* splitAgglomerateInMapping(
-      activeMapping,
-      latestSourceAgglomerateId,
-      [sourceInfo.unmappedId],
-      volumeTracingId,
-      annotationVersion,
-      autoUpdateAgglomerateTrees,
-    );
-
-    if (splitMappingInfo == null) {
-      console.error("Failed to split mapping in proofreading action. Aborting...");
-      return;
-    }
-    const { mappingWithSplitApplied } = splitMappingInfo;
-
-    yield* put(
-      setMappingDataAction(
-        volumeTracingId,
-        mappingWithSplitApplied,
-        // As these split actions were already sent to the server, mappingWithSplitApplied is stored on the server already.
-        true,
-      ),
-    );
-
-    // Now that the local mapping information has changed due to the reloading of the mapping,
-    // reload the agglomerate id info and the mapping.
-    const newInfo = yield* call(
-      reloadMappingAndAggloIds,
-      volumeTracing.tracingId,
-      sourceInfo.unmappedId,
-      targetInfo.unmappedId,
-    );
-    if (!newInfo) {
-      Toast.error(
-        `Could not reload the agglomerate information for proofreading operation segments.`,
-      );
-      return;
-    }
-    activeMapping = newInfo.activeMapping;
-    sourceAgglomerateId = newInfo.sourceAgglomerateId;
-    targetAgglomerateId = newInfo.targetAgglomerateId;
-
-    yield* call(
-      refreshProofreadingSegmentsAndMeshes,
-      volumeTracingId,
-      sourceInfo,
-      targetInfo,
-      sourceAgglomerateId,
-      targetAgglomerateId,
-      ctx,
-    );
-  } finally {
-    if (unsubscribeFromAnnotationMutex) {
-      yield* call(unsubscribeFromAnnotationMutex);
-    }
-  }
-}
-
-function* handleProofreadCutFromNeighbors(action: Action, ctx: OperationContext) {
-  // Actually, action is CutAgglomerateFromNeighborsAction but the
-  // takeEveryUnlessBusy wrapper does not understand this.
-  if (action.type !== "CUT_AGGLOMERATE_FROM_NEIGHBORS") {
-    return;
-  }
-
-  // This action does not depend on the active agglomerate. Instead, it
-  // only depends on the rightclicked agglomerate.
-
-  const allowUpdate = yield* select(mayEditAnnotation);
-  if (!allowUpdate) return;
-
-  const preparation = yield* call(prepareSplitOrMerge, false, ctx);
-  if (!preparation) {
-    return;
-  }
-  const { agglomerateFileMag, getDataValue, volumeTracing, annotationVersion } = preparation;
-  const { tracingId: volumeTracingId } = volumeTracing;
-
-  let idInfos;
-  let targetPosition = null;
-  if (action.position != null) {
-    targetPosition = V3.floor(action.position);
-    idInfos = yield* call(getAgglomerateInfos, preparation.getMappedAndUnmapped, [targetPosition]);
-  } else {
-    if (action.agglomerateId == null || action.segmentId == null) {
-      Toast.warning(MISSING_INFORMATION_WARNING);
-      console.log("Some fields were null:", {
-        agglomerateId: action.agglomerateId,
-        segmentId: action.segmentId,
-      });
-      return;
-    }
-    idInfos = [{ agglomerateId: action.agglomerateId, unmappedId: action.segmentId }];
-
-    targetPosition = yield* call(getPositionForSegmentId, volumeTracing, action.segmentId);
-  }
-  if (!idInfos) {
-    return;
-  }
-  let targetAgglomerateId = idInfos[0].agglomerateId;
-  const targetSegmentId = idInfos[0].unmappedId;
-
-  /* Send the respective split/merge update action to the backend (by pushing to the save queue
-     and saving immediately) */
-  const updateActions: UpdateActionWithoutIsolationRequirement[] = [];
-
-  const { didCancel, neighborInfo } = yield* call(
-    performCutFromNeighbors,
-    targetAgglomerateId,
-    targetSegmentId,
-    targetPosition,
-    agglomerateFileMag,
-    volumeTracingId,
-    annotationVersion,
-    action.tree,
-    updateActions,
-  );
-  if (didCancel || updateActions.length === 0) {
-    return;
-  }
-
-  // Push same info twice as the interface currently, required source & target but cut from all neighbors only has one.
-  yield* call(pushPendingProofreadingOperationInfo, volumeTracingId, idInfos[0], idInfos[0]);
-
-  yield* put(pushSaveQueueTransaction(updateActions));
-  const unsubscribeFromAnnotationMutex = yield* call(
-    subscribeToAnnotationMutexInLiveCollab,
-    "Proofreading Split From All Neighbors",
-  );
-  try {
-    const dummyInfo = idInfos[0];
-    const postProcessResult = yield* call(
-      syncAndUpdatePostProcessingInfo,
-      dummyInfo,
-      dummyInfo,
-      ctx,
-    );
-    if (!postProcessResult) return;
-    const targetAgglomerateIdBeforeSplit = postProcessResult.sourceInfo.agglomerateId;
-
-    // Get active mapping after saving and thus syncing with the backend as this might have changed the mapping.
-    const activeMapping = yield* select(
-      (store) => store.temporaryConfiguration.activeMappingByLayer[volumeTracing.tracingId],
-    );
-
-    // The agglomerateId might have changed due to syncing with the server. Reload it from the mapping.
-    targetAgglomerateId = lookupAgglomerateId(activeMapping, targetSegmentId, targetAgglomerateId);
-
-    const newAnnotationVersion = yield* select((state) => state.annotation.version);
-    // Now that the changes are saved, we can split the mapping locally (because it requires
-    // communication with the back-end).
-    const autoUpdateAgglomerateTrees = true;
-    const splitMappingInfo = yield* splitAgglomerateInMapping(
-      activeMapping,
-      targetAgglomerateId,
-      [targetSegmentId],
-      volumeTracingId,
-      newAnnotationVersion,
-      autoUpdateAgglomerateTrees,
-    );
-
-    if (splitMappingInfo == null) {
-      console.error("Failed to split mapping in cut from all neighbors. Aborting...");
-      return;
-    }
-    const { mappingWithSplitApplied } = splitMappingInfo;
-
-    yield* put(
-      setMappingDataAction(
-        volumeTracingId,
-        mappingWithSplitApplied,
-        // As these split actions were already sent to the server, mappingWithSplitApplied is stored on the server already.
-        true,
-      ),
-    );
-
-    const [newTargetAgglomerateId, ...newNeighborAgglomerateIds] = yield* all([
-      call(getDataValue, targetPosition, mappingWithSplitApplied),
-      ...neighborInfo.neighbors.map((neighbor) =>
-        call(getDataValue, neighbor.position, mappingWithSplitApplied),
-      ),
-    ]);
-
-    /* Ensure segment items exist for affected segments and reload affected meshes */
-    const refreshInfos = [
-      {
-        oldAgglomerateId: targetAgglomerateIdBeforeSplit,
-        newAgglomerateId: newTargetAgglomerateId,
-        nodePosition: targetPosition,
-      },
-      ...neighborInfo.neighbors.map((neighbor, idx) => ({
-        oldAgglomerateId: targetAgglomerateIdBeforeSplit,
-        newAgglomerateId: newNeighborAgglomerateIds[idx],
-        nodePosition: neighbor.position,
-      })),
-    ];
-    yield* call(refreshAffectedSegmentItems, volumeTracingId, refreshInfos);
-
-    yield* call(syncWithBackend, ctx);
-
-    // Refreshing the meshes might take a while and won't block the saga
-    // here.
-    yield* spawnUntilCanceled(maybeRefreshAffectedMeshes, volumeTracingId, refreshInfos);
-  } finally {
-    if (unsubscribeFromAnnotationMutex) {
-      yield* call(unsubscribeFromAnnotationMutex);
-    }
-  }
-}
-
-// Helper functions
-
-type Preparation = {
-  agglomerateFileMag: Vector3;
-  getDataValue: (position: Vector3, overrideMapping?: Mapping | null) => Promise<number>;
-  mapSegmentId: (segmentId: number, overrideMapping?: Mapping | null) => number;
-  getMappedAndUnmapped: (position: Vector3) => Saga<{ agglomerateId: number; unmappedId: number }>;
-  activeMapping: ActiveMappingInfo;
-  volumeTracing: VolumeTracing & { mappingName: string };
-  annotationVersion: number;
-};
-
-export function* prepareSplitOrMerge(
-  isTreeProofreading: boolean,
-  ctx: OperationContext,
-): Saga<Preparation | null> {
-  const volumeTracingLayer = yield* select((state) => getActiveSegmentationTracingLayer(state));
-  const volumeTracing = yield* select((state) => getActiveSegmentationTracing(state));
-  if (volumeTracingLayer == null || volumeTracing == null) {
-    return null;
-  }
-  let { mappingName } = volumeTracing;
-  if (mappingName == null) {
-    return null;
-  }
-
-  const isHdf5MappingEnabled = yield* call(ensureHdf5MappingIsEnabled, volumeTracing.tracingId);
-  if (!isHdf5MappingEnabled) {
-    return null;
-  }
-
-  if (!volumeTracing.hasEditableMapping) {
-    try {
-      mappingName = yield* call(createEditableMapping, ctx);
-    } catch (e) {
-      console.error(e);
-      return null;
-    }
-  }
-
-  const magInfo = getMagInfo(volumeTracingLayer.mags);
-  const currentMag = yield* select((state) => getCurrentMag(state, volumeTracingLayer.name));
-
-  const agglomerateFileMag = isTreeProofreading
-    ? // In case of tree proofreading, the finest mag should be used.
-      magInfo.getFinestMag()
-    : // For non-tree proofreading, the active mag suffices
-      currentMag;
-  if (agglomerateFileMag == null) {
-    return null;
-  }
-  const agglomerateFileZoomstep = magInfo.getIndexByMag(agglomerateFileMag);
-
-  const getUnmappedDataValue = (position: Vector3): Promise<number> => {
-    const { additionalCoordinates } = Store.getState().flycam;
-    return api.data.getDataValue(
-      volumeTracing.tracingId,
-      position,
-      agglomerateFileZoomstep,
-      additionalCoordinates,
-    );
-  };
-
-  console.log("Accessing mapping for proofreading");
-  const mapping = yield* select(
-    (state) =>
-      getMappingInfo(state.temporaryConfiguration.activeMappingByLayer, volumeTracing.tracingId)
-        .mapping,
-  );
-
-  if (mapping == null) {
-    Toast.warning("Mapping is not available, cannot proofread.");
-    return null;
-  }
-
-  const getDataValue = async (
-    position: Vector3,
-    overrideMapping: Mapping | null = null,
-  ): Promise<number> => {
-    const unmappedId = await getUnmappedDataValue(position);
-    return mapSegmentId(unmappedId, overrideMapping);
-  };
-
-  const mapSegmentId = (segmentId: number, overrideMapping: Mapping | null = null): number => {
-    const mappingToAccess = overrideMapping ?? mapping;
-    const mappingWrapper = new NumberLikeMapWrapper(mappingToAccess);
-    const mappedId = mappingWrapper.getAsNumber(segmentId);
-    if (mappedId == null) {
-      // It could happen that the user tries to perform a proofreading operation
-      // that involves an id for which the mapped id wasn't fetched yet.
-      // In that case, we currently just throw an error. A toast will appear
-      // that asks the user to retry. If we notice that this happens in production,
-      // we can think about a better way to handle this.
-      throw new SoftError(
-        `Could not map id ${segmentId}. The mapped partner might not be known yet. Please retry.`,
-      );
-    }
-    return mappedId;
-  };
-
-  const getMappedAndUnmapped = function* (position: Vector3) {
-    const unmappedId = yield* call(getUnmappedDataValue, position);
-    let agglomerateId = isNumberMap(mapping)
-      ? mapping.get(unmappedId)
-      : // TODO: Proper 64 bit support (#6921)
-        Number(mapping.get(BigInt(unmappedId)));
-
-    if (agglomerateId == null) {
-      const fetchedEntries = yield* call(
-        getAgglomeratesForSegmentIds,
-        volumeTracing.tracingId,
-        volumeTracingLayer,
-        mappingName,
-        new Set([unmappedId]),
-      );
-      agglomerateId = new NumberLikeMapWrapper(fetchedEntries).getAsNumber(unmappedId);
-      if (agglomerateId == null) {
-        throw new SoftError(
-          `Could not map id ${unmappedId} at position ${position}. The mapped partner might not be known yet. Please retry.`,
-        );
-      }
-    }
-    return { agglomerateId, unmappedId };
-  };
-
-  const activeMapping = yield* select(
-    (store) => store.temporaryConfiguration.activeMappingByLayer[volumeTracing.tracingId],
-  );
-
-  if (activeMapping.mapping == null) {
-    Toast.error("Active mapping is not available, cannot proofread.");
-    return null;
-  }
-
-  // Getting latest annotation version as it might have changed due to e.g. making the mapping editable.
-  const annotationVersion = yield* select((state) => state.annotation.version);
-
-  return {
-    agglomerateFileMag,
-    getDataValue,
-    getMappedAndUnmapped,
-    mapSegmentId,
-    activeMapping,
-    volumeTracing: { ...volumeTracing, mappingName },
-    annotationVersion,
-  };
-}
-
-function* reloadMappingAndAggloIds(
-  tracingId: string,
-  unmappedSourceId: number,
-  unmappedTargetId: number,
-) {
-  const activeMapping = yield* select(
-    (store) => store.temporaryConfiguration.activeMappingByLayer[tracingId],
-  );
-
-  if (activeMapping.mapping != null) {
-    const mappingWrapper = new NumberLikeMapWrapper(activeMapping.mapping);
-
-    const maybeSourceAgglomerateId = mappingWrapper.getAsNumber(unmappedSourceId);
-    const maybeTargetAgglomerateId = mappingWrapper.getAsNumber(unmappedTargetId);
-
-    if (maybeSourceAgglomerateId != null && maybeTargetAgglomerateId != null) {
-      return {
-        sourceAgglomerateId: maybeSourceAgglomerateId,
-        targetAgglomerateId: maybeTargetAgglomerateId,
-        activeMapping,
-      };
-    }
-
-    // As an additional safety net we look up the IDs again. Actually,
-    // this should not happen in production.
-    const segmentsToRequest = new Set([unmappedSourceId, unmappedTargetId]);
-    const tracingStoreUrl = yield* select((state) => state.annotation.tracingStore.url);
-    const annotationVersion = yield* select((state) => state.annotation.version);
-    const annotationId = yield* select((state) => state.annotation.annotationId);
-
-    const missingMappingInfo = yield* call(
-      getAgglomeratesForSegmentsFromTracingstore,
-      tracingStoreUrl,
-      tracingId,
-      segmentsToRequest,
-      annotationId,
-      annotationVersion,
-    );
-    const missingMappingInfoWrapper = new NumberLikeMapWrapper(missingMappingInfo);
-    const maybeSourceAgglomerateIdFromServer =
-      missingMappingInfoWrapper.getAsNumber(unmappedSourceId);
-    const maybeTargetAgglomerateIdFromServer =
-      missingMappingInfoWrapper.getAsNumber(unmappedTargetId);
-    if (maybeSourceAgglomerateIdFromServer == null || maybeTargetAgglomerateIdFromServer == null) {
-      throw new SoftError(
-        `Could not look up agglomerate id for unmapped segment id ${unmappedSourceId} or ${unmappedTargetId}.`,
-      );
-    }
-    return {
-      sourceAgglomerateId: maybeSourceAgglomerateIdFromServer,
-      targetAgglomerateId: maybeTargetAgglomerateIdFromServer,
-      activeMapping,
-    };
-  }
-}
-
-function* getAgglomerateInfos(
-  getMappedAndUnmapped: (position: Vector3) => Saga<{ agglomerateId: number; unmappedId: number }>,
-  positions: Vector3[],
-): Saga<Array<IdInfoWithoutPosition> | null> {
-  try {
-    const idInfos = yield* all(positions.map((pos) => call(getMappedAndUnmapped, pos)));
-    if (idInfos.find((idInfo) => idInfo.agglomerateId === 0 || idInfo.unmappedId === 0) != null) {
-      Toast.warning(
-        "One of the selected segments has the id 0 which is the background. Cannot merge/split.",
-      );
-      console.warn("At least one id was zero:", idInfos);
-      return null;
-    }
-    return idInfos;
-  } catch (exception) {
-    Toast.error("Cannot perform proofreading operation. Please retry. See console for details.");
-    console.error(exception);
-    return null;
-  }
-}
-
-export function* refreshAffectedSegmentItems(
-  layerName: string,
-  items: Array<{
-    oldAgglomerateId?: number;
-    newAgglomerateId: number;
-    nodePosition: Vector3;
-  }>,
-) {
-  // Segmentations with more than 3 dimensions are currently not compatible
-  // with proofreading. Once such datasets appear, this parameter needs to be
-  // adapted.
-  const additionalCoordinates = undefined;
-  // Remove old segments which are no longer present.
-  const outdatedIds = uniq(items.map((item) => item.oldAgglomerateId)).filter((id) => id != null);
-  const itemsToAddOrUpdate = uniqBy(items, (item) => item.newAgglomerateId);
-  const removedIds = new Set(outdatedIds).difference(
-    new Set(itemsToAddOrUpdate.map((item) => item.newAgglomerateId)),
-  );
-  const removeEffects = [...removedIds].map((id) => put(removeSegmentAction(id, layerName)));
-  yield* all(removeEffects);
-
-  const ensureSegmentItemEffects = uniqBy(items, (item) => item.newAgglomerateId).map((item) =>
-    call(
-      ensureSegmentItem,
-      layerName,
-      Number(item.newAgglomerateId),
-      item.nodePosition,
-      additionalCoordinates,
-    ),
-  );
-  // By using `all`, we avoid problems which can occur when running too many
-  // call effects in a for loop. Also see https://github.com/redux-saga/redux-saga/issues/1592.
-  yield* all(ensureSegmentItemEffects);
-}
-
-export function* shouldReloadMeshesAfterProofreadAction(
-  layerName: string,
-  oldAgglomerateIds: number[],
-): Saga<boolean> {
-  const autoRenderMeshInProofreading = yield* select(
-    (state) => state.userConfiguration.autoRenderMeshInProofreading,
-  );
-  if (autoRenderMeshInProofreading) {
-    return true;
-  }
-  const hasAnyInvolvedMeshLoaded = yield* select((state) =>
-    oldAgglomerateIds.some((id) => isMeshLoaded(state, id, layerName)),
-  );
-  return hasAnyInvolvedMeshLoaded;
-}
-
-// Display properties of a mesh that should survive a reload.
-export type PreservedMeshDisplayProps = {
-  opacity?: number;
-  isVisible?: boolean;
-};
-
-// Capture the current opacity and visibility of the given old agglomerates' meshes, keyed by
-// agglomerate id, so that reloaded meshes can keep the user-chosen opacity and visibility.
-// Duplicate and nullish ids are ignored, so callers can pass the raw oldAgglomerateId of every
-// item even when several share the same id (e.g. a split produces two pieces from the same
-// original agglomerate). Callers must invoke this before any old mesh is removed, otherwise the
-// original mesh display properties can no longer be read.
-export function* getMeshDisplayPropsByOldAgglomerateId(
-  layerName: string,
-  oldAgglomerateIds: Iterable<number | null | undefined>,
-  additionalCoordinates: AdditionalCoordinate[] | null | undefined,
-): Saga<Map<number, PreservedMeshDisplayProps>> {
-  return yield* select((state) => {
-    const displayPropsByAgglomerateId = new Map<number, PreservedMeshDisplayProps>();
-    for (const oldAgglomerateId of oldAgglomerateIds) {
-      if (oldAgglomerateId != null && !displayPropsByAgglomerateId.has(oldAgglomerateId)) {
-        const meshInfo = getMeshInfoForSegment(
-          state,
-          additionalCoordinates || null,
-          layerName,
-          Number(oldAgglomerateId),
-        );
-        if (meshInfo != null) {
-          displayPropsByAgglomerateId.set(oldAgglomerateId, {
-            opacity: meshInfo.opacity,
-            isVisible: meshInfo.isVisible,
-          });
-        }
-      }
-    }
-    return displayPropsByAgglomerateId;
-  });
-}
-
-function* maybeRefreshAffectedMeshes(
-  layerName: string,
-  items: Array<{
-    oldAgglomerateId?: number;
-    newAgglomerateId: number;
-    nodePosition: Vector3;
-    opacity?: number; // see refreshAffectedMeshes below.
-  }>,
-) {
-  const shouldDoMeshRefreshing = yield* call(shouldReloadMeshesAfterProofreadAction, layerName, [
-    ...items.map((i) => i.oldAgglomerateId).filter((id) => id != null),
-  ]);
-  if (shouldDoMeshRefreshing) {
-    // Refreshing the meshes might take a while and won't block the saga
-    // here.
-    yield* spawnUntilCanceled(refreshAffectedMeshes, layerName, items);
-  }
-}
-
-export function* refreshAffectedMeshes(
-  layerName: string,
-  items: Array<{
-    oldAgglomerateId?: number;
-    newAgglomerateId: number;
-    nodePosition: Vector3;
-    // Opacity and visibility to apply to the reloaded mesh. If unset, the values of the old
-    // mesh (oldAgglomerateId) are used before its removal (see below).
-    opacity?: number;
-    isVisible?: boolean;
-  }>,
-) {
-  // ATTENTION: This saga should usually be called with `spawnUntilCanceled` to avoid that the user
-  // is blocked (via takeEveryUnlessBusy) while the meshes are refreshed.
-
-  // Segmentations with more than 3 dimensions are currently not compatible
-  // with proofreading. Once such datasets appear, this parameter needs to be
-  // adapted.
-  const additionalCoordinates = undefined;
-
-  // Capture the opacity and visibility of all old meshes up front, i.e. before any of them are
-  // removed below, so that reloaded meshes keep the user-chosen opacity and visibility. This must
-  // happen before the removal loop because removing one item's old mesh must not prevent another
-  // item from reading the original properties.
-  const displayPropsByOldAgglomerateId = yield* call(
-    getMeshDisplayPropsByOldAgglomerateId,
-    layerName,
-    items.map((item) => item.oldAgglomerateId),
-    additionalCoordinates,
-  );
-
-  // Remember which meshes were removed in this saga
-  // and which were fetched again to avoid doing redundant work.
-  const removedIds = new Set();
-  const newlyLoadedIds = new Set();
-  const meshLoadingEffects = [];
-  for (const item of items) {
-    // Opacity and visibility are either passed in explicitly (e.g. by the rebasing saga, which
-    // removes the old mesh before this saga runs) or taken from the old mesh captured above.
-    const oldDisplayProps =
-      item.oldAgglomerateId != null
-        ? displayPropsByOldAgglomerateId.get(item.oldAgglomerateId)
-        : undefined;
-    const opacity = item.opacity ?? oldDisplayProps?.opacity;
-    const isVisible = item.isVisible ?? oldDisplayProps?.isVisible;
-    // Remove old agglomerate mesh(es) and load updated agglomerate mesh(es)
-    if (item.oldAgglomerateId && !removedIds.has(item.oldAgglomerateId)) {
-      yield* put(removeMeshAction(layerName, Number(item.oldAgglomerateId)));
-      removedIds.add(item.oldAgglomerateId);
-    }
-    if (!newlyLoadedIds.has(item.newAgglomerateId)) {
-      meshLoadingEffects.push(function* load() {
-        yield* call(
-          loadCoarseMesh,
-          layerName,
-          Number(item.newAgglomerateId),
-          item.nodePosition,
-          additionalCoordinates,
-          opacity,
-          isVisible,
-        );
-      });
-      newlyLoadedIds.add(item.newAgglomerateId);
-    }
-  }
-  // Do all mesh loadings in parallel for more speed.
-  yield* call(
-    processTaskWithPool,
-    meshLoadingEffects,
-    Constants.PARALLEL_PRECOMPUTED_MESH_LOADING_COUNT,
-  );
-}
-
-function getDeleteEdgeActionForEdgePositions(
-  sourceTree: Tree,
-  edge: { position1: Vector3; position2: Vector3 },
-) {
-  let firstNodeId;
-  let secondNodeId;
-  for (const node of sourceTree.nodes.values()) {
-    if (isEqual(node.untransformedPosition, edge.position1)) {
-      firstNodeId = node.id;
-    } else if (isEqual(node.untransformedPosition, edge.position2)) {
-      secondNodeId = node.id;
-    }
-    if (firstNodeId && secondNodeId) {
-      break;
-    }
-  }
-
-  if (!firstNodeId || !secondNodeId) {
-    Toast.warning(
-      `Unable to find all nodes for positions ${!firstNodeId ? edge.position1 : null}${
-        !secondNodeId ? [", ", edge.position2] : null
-      } in ${sourceTree.name}.`,
-    );
-    return null;
-  }
-  return { firstNodeId, secondNodeId };
-}
-
-function* getPositionForSegmentId(volumeTracing: VolumeTracing, segmentId: number): Saga<Vector3> {
-  const dataset = yield* select((state) => state.dataset);
-  const dataStoreUrl = yield* select((state) => state.dataset.dataStore.url);
-  const editableMapping = yield* select((state) =>
-    getEditableMappingForVolumeTracingId(state, volumeTracing.tracingId),
-  );
-  if (volumeTracing.fallbackLayer == null || editableMapping == null) {
-    // Should not happen in proofreading.
-    throw new Error("Could not find fallback layer or editable mapping.");
-  }
-  const position = yield* call(
-    getPositionForSegmentInAgglomerate,
-    dataStoreUrl,
-    dataset.id,
-    volumeTracing.fallbackLayer,
-    editableMapping.baseMappingName,
-    segmentId,
-  );
-  return position;
-}
-
-function getSegmentIdsThatMapToAgglomerate(
-  activeMapping: ActiveMappingInfo,
-  sourceAgglomerateId: number,
-) {
-  // Obtain all segment ids that map to sourceAgglomerateId
-  const mappingEntries = Array.from(activeMapping.mapping as NumberLikeMap);
-
-  const adaptToType = getAdaptToTypeFunction(activeMapping.mapping);
-
-  // If the mapping contains BigInts, we need a BigInt for the filtering
-  const comparableSourceAgglomerateId = adaptToType(sourceAgglomerateId);
-  return mappingEntries
-    .filter(([_segmentId, agglomerateId]) => agglomerateId === comparableSourceAgglomerateId)
-    .map(([segmentId, _agglomerateId]) => segmentId);
-}
-
-type SplitAgglomeratesInMappingResult = {
-  mappingWithSplitApplied: Mapping;
-  oldAgglomerateIds: Set<number>;
-  newAgglomerateIds: Set<number>;
-  newToOldAgglomerateIds: Map<number, number>;
-};
-
-/**
- * Re-builds the local mapping to reflect one or more server-side agglomerate splits by re-requesting the
- * affected segments' new agglomerate ids and returning the updated mapping plus the involved old/new ids.
- *
- * Affected segment IDs are found by scanning the local mapping for ids that map to the provided agglomerate
- * IDs. Additionally, all keys of the `segmentIdToOldAgglomerateId` parameter are also considered to be
- * "affected". `segmentIdToOldAgglomerateId` doesn't need to contain *all* segment ids that map to a common
- * agglomerate ID. Instead, a segment ID only needs to be contained if the caller needs the ID to be part of
- * the returned mapping (`addAdditionalSegmentsToMapping` needs to be set to true, too).
- *
- * This is the general (plural) implementation which can split several agglomerates at once. For the common
- * case of splitting a single agglomerate, prefer the {@link splitAgglomerateInMapping} wrapper.
- *
- * @param activeMapping - Current mapping info which is copied and mutated to reflect the reassigned segments
- *   before returning the new mapping as mappingWithSplitApplied.
- * @param segmentIdToOldAgglomerateId - (Some¹) segment IDs involved in the split action, mapped to its pre-split
- *   agglomerate id. Must contain at least one segment of the split agglomerate. Segments may be passed even if
- *   they are not in the local mapping (e.g. while forwarded foreign splits during rebasing).
- *   ¹ See paragraph about "affected segment ids" above for more details about the impact of passed IDs.
- * @param volumeTracingId - Id of the volume tracing whose (editable) mapping is updated.
- * @param version - Annotation version at which to look up the post-split agglomerate ids.
- * @param syncAgglomerateTrees - If true, sync the skeleton trees to the split.
- * @param addAdditionalSegmentsToMapping - Whether to also write the passed segments into the returned
- *   mapping which are currently not part of the activeMapping.
- *  (needed e.g. by partitioned min-cut; forwarded foreign splits keep them out).
- * @returns `undefined` if nothing to split and the input mapping is `null`; otherwise `{ mappingWithSplitApplied`
- *   (rebuilt full mapping)`, oldAgglomerateIds` (ids before the split)`, newAgglomerateIds` (new ids produced
- *   by the split)`, newToOldAgglomerateIds` (each new id → the old id it came from, e.g. for carrying over
- *   mesh opacity)`}`.
- */
-export function* splitAgglomeratesInMapping(
-  activeMapping: ActiveMappingInfo,
-  segmentIdToOldAgglomerateId: Map<number, number>,
-  volumeTracingId: string,
-  version: number,
-  syncAgglomerateTrees: boolean,
-  addAdditionalSegmentsToMapping = false,
-): Saga<SplitAgglomeratesInMappingResult | undefined> {
-  // The old agglomerate ids are all agglomerates the passed segments belonged to before the split.
-  const oldAgglomerateIds = new Set<number>(segmentIdToOldAgglomerateId.values());
-  // Re-map the local segments of EVERY involved old agglomerate, so that segments of those
-  // agglomerates are not left in an outdated state.
-  const localSegmentIds = new Set(
-    Array.from(oldAgglomerateIds).flatMap((oldAgglomerateId) =>
-      getSegmentIdsThatMapToAgglomerate(activeMapping, oldAgglomerateId),
-    ),
-  );
-  const splitSegmentIds = localSegmentIds.union(new Set(segmentIdToOldAgglomerateId.keys()));
-  const annotationId = yield* select((state) => state.annotation.annotationId);
-  const tracingStoreUrl = yield* select((state) => state.annotation.tracingStore.url);
-  // Ask the server to map the (split) segment ids. This creates a partial mapping
-  // that only contains these ids.
-  const mappingBeforeSplit = activeMapping.mapping;
-  if (splitSegmentIds.size === 0 || mappingBeforeSplit == null) {
-    return mappingBeforeSplit != null
-      ? {
-          mappingWithSplitApplied: mappingBeforeSplit,
-          newAgglomerateIds: new Set(),
-          oldAgglomerateIds: new Set(),
-          newToOldAgglomerateIds: new Map(),
-        }
-      : undefined;
-  }
-  const requestedMappingInfo = yield* call(
-    getAgglomeratesForSegmentsFromTracingstore,
-    tracingStoreUrl,
-    volumeTracingId,
-    splitSegmentIds,
-    annotationId,
-    version,
-  );
-  const newToOldAgglomerateIds = new Map<number, number>();
-
-  const knownAndAdditionalSegmentIds = new Set<number>(
-    Array.from(mappingBeforeSplit.keys() ?? [], Number),
-  ).union(new Set(segmentIdToOldAgglomerateId.keys()));
-
-  const requestedMappingInfoWrapped = new NumberLikeMapWrapper(requestedMappingInfo);
-  const mappingBeforeSplitWrapped = new NumberLikeMapWrapper(mappingBeforeSplit);
-
-  // Create a new mapping which is equal to the old one with the difference that
-  // ids requested from the backend are mapped to their new target agglomerate ids.
-  // If addAdditionalSegmentsToMapping = true all requested segment ids are included
-  // in the mapping. Even those not present before but requested. In parallel the
-  // newToOldAgglomerateIds map is built and then used on the caller site to maintain
-  // mesh opacity and visibility between reloading the meshes.
-  const mappingWithSplitApplied = new Map();
-  knownAndAdditionalSegmentIds.forEach((segmentId) => {
-    const mappedId = requestedMappingInfoWrapped.getAsNumber(segmentId);
-    if (mappedId != null) {
-      const prevAgglomerateId =
-        segmentIdToOldAgglomerateId.get(segmentId) ??
-        mappingBeforeSplitWrapped.getAsNumber(segmentId);
-      newToOldAgglomerateIds.set(mappedId, Number(prevAgglomerateId));
-      if (mappingBeforeSplitWrapped.has(segmentId) || addAdditionalSegmentsToMapping) {
-        mappingWithSplitApplied.set(segmentId, mappedId);
-      }
-    } else {
-      if (splitSegmentIds.has(segmentId)) {
-        console.error(
-          "Got a agglomerate id mapping reply from the server which did not include the new agglomerate id of a requested segment.",
-          `${segmentId} was not present in ${requestedMappingInfo.entries()}.`,
-        );
-      }
-      // The segment was not requested / not included in the reply, but already present in the previous mapping. Just copy it over.
-      const unchangedAgglomerateId = mappingBeforeSplitWrapped.getAsNumber(segmentId);
-      mappingWithSplitApplied.set(segmentId, unchangedAgglomerateId);
-    }
-  });
-
-  const newAgglomerateIds = new Set(newToOldAgglomerateIds.keys());
-  if (syncAgglomerateTrees && activeMapping.mappingName) {
-    yield* call(
-      syncAgglomerateTreesAfterSplitAction,
-      Array.from(newAgglomerateIds),
-      Array.from(oldAgglomerateIds),
-      volumeTracingId,
-    );
-  }
-
-  return {
-    mappingWithSplitApplied: mappingWithSplitApplied as Mapping,
-    oldAgglomerateIds,
-    newAgglomerateIds,
-    newToOldAgglomerateIds,
-  };
-}
-
-/**
- * Lightweight wrapper around {@link splitAgglomeratesInMapping} for the common case of splitting a
- * single agglomerate. It tags every affected segment with the same `agglomerateId` and delegates to
- * the plural implementation.
- *
- * @param agglomerateId - The agglomerate id that was split.
- * @param affectedSegmentIds - Segment ids known to have belonged to `agglomerateId` before the split.
- *   At least one must be provided so the split can be resolved even when the agglomerate has no
- *   segments in the local mapping. Segments that are not part of the local mapping are only written
- *   into the returned mapping when `addAdditionalSegmentsToMapping` is set.
- *   The parameter doesn't need to contain *all* segment ids that map to a common agglomerate ID
- *  (see referenced function above for more information).
- *
- * @see splitAgglomeratesInMapping for the remaining parameters and the return value.
- */
-export function* splitAgglomerateInMapping(
-  activeMapping: ActiveMappingInfo,
-  agglomerateId: number,
-  affectedSegmentIds: Iterable<number>,
-  volumeTracingId: string,
-  version: number,
-  syncAgglomerateTrees: boolean,
-  addAdditionalSegmentsToMapping = false,
-): Saga<SplitAgglomeratesInMappingResult | undefined> {
-  const segmentIdToOldAgglomerateId = new Map(
-    Array.from(affectedSegmentIds, (segmentId) => [segmentId, agglomerateId] as [number, number]),
-  );
-  return yield* splitAgglomeratesInMapping(
-    activeMapping,
-    segmentIdToOldAgglomerateId,
-    volumeTracingId,
-    version,
-    syncAgglomerateTrees,
-    addAdditionalSegmentsToMapping,
-  );
-}
-
-function* mergeAgglomeratesInMapping(
-  activeMapping: ActiveMappingInfo,
-  sourceAgglomerateId: number,
-  targetAgglomerateId: number,
-): Saga<Mapping> {
-  const adaptToType = getAdaptToTypeFunction(activeMapping.mapping);
-
-  const typedTargetAgglomerateId = adaptToType(targetAgglomerateId);
-  const typedSourceAgglomerateId = adaptToType(sourceAgglomerateId);
-  return new Map(
-    Array.from(activeMapping.mapping as NumberLikeMap, ([key, value]) =>
-      value === typedTargetAgglomerateId ? [key, typedSourceAgglomerateId] : [key, value],
-    ),
-  ) as Mapping;
-}
-
-export function* updateMappingWithMerge(
-  volumeTracingId: string,
-  activeMapping: ActiveMappingInfo,
-  sourceAgglomerateId: number,
-  targetAgglomerateId: number,
-  // Must be true if we know that the updated mapping info was already saved on the server.
-  isVersionStoredOnServer: boolean,
-) {
-  const mergedMapping = yield* call(
-    mergeAgglomeratesInMapping,
-    activeMapping,
-    sourceAgglomerateId,
-    targetAgglomerateId,
-  );
-  // Even when the merge was a no-op, we still dispatch setMappingDataAction here, so the activation completes correctly (see the finishMappingActivation saga).
-  yield* put(setMappingDataAction(volumeTracingId, mergedMapping, isVersionStoredOnServer));
-}
-
-type IdInfo = { agglomerateId: number; unmappedId: number; position: Vector3 };
-type IdInfoOpt = { agglomerateId: number; unmappedId: number; position: Vector3 | undefined };
-type IdInfoWithoutPosition = { agglomerateId: number; unmappedId: number };
-
-type GatheredInfos =
-  | {
-      type: "PROOFREAD_MERGE";
-      infos: [IdInfo, IdInfoOpt];
-    }
-  | {
-      type: "MIN_CUT_AGGLOMERATE";
-      infos: [IdInfo, IdInfo];
-    };
-
-function* gatherInfoForOperation(
-  action: ProofreadMergeAction | MinCutAgglomerateWithPositionAction,
-  preparation: Preparation,
-): Saga<GatheredInfos | null> {
-  const { volumeTracing } = preparation;
-  const { tracingId: volumeTracingId, activeCellId } = volumeTracing;
-  const activeUnmappedSegmentId = yield* select((state) =>
-    getActiveUnmappedSegmentId(state, volumeTracing),
-  );
-  if (activeCellId === 0) {
-    console.warn("[Proofreading] Cannot execute operation because active segment id is 0");
-    return null;
-  }
-
-  const segments = yield* select((store) => getSegmentsForLayer(store, volumeTracingId));
-  const activeSegment = segments.getNullable(activeCellId);
-  if (activeSegment == null) {
-    console.warn("[Proofreading] Cannot execute operation because no active segment item exists");
-    return null;
-  }
-  const activeSegmentPositionFloat = activeSegment.anchorPosition;
-  if (activeSegmentPositionFloat == null) {
-    console.warn("[Proofreading] Cannot execute operation because active segment has no position");
-    return null;
-  }
-
-  const activeSegmentPosition = V3.floor(activeSegmentPositionFloat);
-
-  let sourcePosition: Vector3 | undefined;
-  let targetPosition: Vector3 | undefined;
-
-  if (action.position) {
-    // The action was triggered via a data viewport (not 3D). In this case,
-    // the active segment's position can be used as a source.
-    if (activeUnmappedSegmentId != null) {
-      // The user has selected a supervoxel in the 3D viewport and then clicked
-      // in a data viewport to select the second merge partner. However, this mix
-      // is currently not supported.
-      Toast.warning(MISSING_INFORMATION_WARNING);
-      return null;
-    }
-    sourcePosition = activeSegmentPosition;
-    targetPosition = V3.floor(action.position);
-    const idInfos = yield* call(getAgglomerateInfos, preparation.getMappedAndUnmapped, [
-      sourcePosition,
-      targetPosition,
-    ]);
-    if (idInfos == null) {
-      console.warn(
-        "[Proofreading] Cannot execute operation because agglomerate infos couldn't be determined for source and target position.",
-      );
-      return null;
-    }
-    const [idInfo1, idInfo2] = idInfos;
-    return {
-      type: action.type,
-      infos: [
-        { ...idInfo1, position: sourcePosition },
-        { ...idInfo2, position: targetPosition },
-      ],
-    };
-  }
-
-  // The action was triggered in the 3D viewport. In this case, we don't have
-  // a mouse position and also the active segment position isn't necessarily
-  // a position of the clicked supervoxel.
-  if (
-    action.agglomerateId == null ||
-    activeCellId == null ||
-    activeUnmappedSegmentId == null ||
-    action.segmentId == null
-  ) {
-    Toast.warning(MISSING_INFORMATION_WARNING);
-    console.log("Some fields were null:", {
-      agglomerateId: action.agglomerateId,
-      activeCellId,
-      activeUnmappedSegmentId,
-      segmentId: action.segmentId,
-    });
-    return null;
-  }
-  const targetSegmentId = action.segmentId;
-
-  if (action.type === "PROOFREAD_MERGE") {
-    const idInfos: [IdInfo, IdInfoOpt] = [
-      {
-        agglomerateId: activeCellId,
-        unmappedId: activeUnmappedSegmentId,
-        position: activeSegmentPosition,
-      },
-      {
-        agglomerateId: action.agglomerateId,
-        unmappedId: action.segmentId,
-        // Strictly speaking, we do not have a valid position for the target segment
-        // here. *After* the merge, the source position would also be valid for
-        // the target, but we leave that logic to the caller.
-        position: undefined,
-      },
-    ];
-
-    return {
-      type: action.type,
-      infos: idInfos,
-    };
-  }
-  // When splitting two segments, we don't really have reliable positions at hand.
-  // For the source position, we cannot rely on the active segment position, because
-  // the active supervoxel doesn't necessarily match the last click position within
-  // the data viewports.
-  // For the target position, we also don't have reliable information available.
-  [sourcePosition, targetPosition] = yield* all([
-    call(getPositionForSegmentId, volumeTracing, activeUnmappedSegmentId),
-    call(getPositionForSegmentId, volumeTracing, targetSegmentId),
-  ]);
-
-  const idInfos: [IdInfo, IdInfo] = [
-    { agglomerateId: activeCellId, unmappedId: activeUnmappedSegmentId, position: sourcePosition },
-    { agglomerateId: action.agglomerateId, unmappedId: action.segmentId, position: targetPosition },
-  ];
-
-  return {
-    type: action.type,
-    infos: idInfos,
-  };
 }
